Mve6 sprite display (#37)
* debut

* debut calculs sprites

* aaarrrgghhhh

* fix to compile

* texture can now be displayed for sprites

* new sprite texture

* new texture + only non-black colors are displayed

* normed

* define

* ajout sprites burning barrel

* retry

* ajout sprites burning barrel et lamp

* is_viewed for sprites

* exploding barrel sprites

* sprites are displayed correctly

* sprites folder

* sprites can now have several textures

* burning_barrel et metallic_lamp corriges
¨

* sprites has several textures in an array

* fix segfaults with empty sprite texture

* sprite animation

* warning if there are too many textures

* add help option

* add help option

* crossable and uncrossable sprite

* sprites réparés

* dark attribute

* sprites finished

* is_dark boolean

* all sprites

* progressive darkness

* fix darkness on ceil and floor

* clean some buid config code

* add security for division by 0 in the darkness color

* fix mandatory darkness

* adjust the help message for darkness

* update backroom_textures

* update backroom_texture wall 2

* update backroom_texture wall 2

* update invalid_rgb and custom very_big_map1

* custom maps + distance modified

* can change the dark_dist

* add profiler option in the Makefile

* fix pixel lines ceil + fix gigachad texture

* fix wall angle crossing

* adapt tester for pr

* fix conversations

* fix invalid_file for sprite (crash)

* fix too big map

* move the map test.cub

---------

+++ includes/cub3d_loop.h
@@ -6,7 +6,7 @@
 /*   By: rbroque <rbroque@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/26 17:14:28 by lechon            #+#    #+#             */
-/*   Updated: 2023/11/09 13:16:35 by rbroque          ###   ########.fr       */
+/*   Updated: 2023/11/15 21:25:05 by rbroque          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -87,8 +87,8 @@ void			translate_frontback(t_map *const map,
 
 	// is_crossable.c
 
-bool			is_crossable(const t_map *const map, const size_t x,
-					const size_t y);
+bool			is_crossable(const t_map *const map, const double x,
+					const double y);
 
 	// get_translated_pos.c
